Understanding and Visualizing Data with Python
Understanding and Visualizing Data with Python Course

Getting started in data basics and statistics. Creating data visualizations and numerical summaries for univariate and multivariate data using Python programming language. Understanding the concept of a sampling distribution, and how estimation of the variance of that distribution plays a critical role in making statements about populations.